+++ packages/nimble-components/src/wafer-map/index.ts
@@ -287,8 +287,10 @@ export class WaferMap<
         renderConfig: RenderConfig,
         dieDimensions: Dimensions,
         transform: ZoomTransform,
+        dieLabelsHidden: boolean,
         columnIndices: Int32Array,
-        rowIndices: Int32Array
+        rowIndices: Int32Array,
+        values: Float64Array
     } {
         const canvasDimensions = {
             width: this.canvasWidth ?? 0,
@@ -301,31 +303,42 @@ export class WaferMap<
             horizontalCoefficient: this.computations.horizontalCoefficient,
             horizontalConstant: this.computations.horizontalConstant,
             verticalConstant: this.computations.verticalConstant,
+            gridMinX: this.computations.gridMinX,
+            gridMaxX: this.computations.gridMaxX,
+            gridMinY: this.computations.gridMinY,
+            gridMaxY: this.computations.gridMaxY,
             labelsFontSize: this.computations.labelsFontSize,
-            colorScale: this.computations.colorScale
+            colorScale: this.computations.colorScale,
+            dieLabelsSuffix: this.dieLabelsSuffix,
+            maxCharacters: this.maxCharacters
         };
         const dieDimensions = this.computations.dieDimensions;
         const transform = this.transform;
+        const dieLabelsHidden = this.dieLabelsHidden;
         if (this.diesTable === undefined) {
             return {
                 canvasDimensions,
                 renderConfig,
                 dieDimensions,
                 transform,
+                dieLabelsHidden,
                 columnIndices: Int32Array.from([]),
-                rowIndices: Int32Array.from([])
+                rowIndices: Int32Array.from([]),
+                values: Float64Array.from([])
             };
         }
         const columnIndices = this.diesTable.getChild('colIndex')!.toArray();
-
         const rowIndices = this.diesTable.getChild('rowIndex')!.toArray();
+        const values = this.diesTable.getChild('value')!.toArray();
         return {
             canvasDimensions,
             renderConfig,
             columnIndices,
             rowIndices,
+            values,
             dieDimensions,
-            transform
+            transform,
+            dieLabelsHidden
         };
     }