+++ packages/angular-workspace/nimble-angular/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Fri, 07 Jun 2024 15:10:38 GMT and should not be manually modified.
+This log was last generated on Fri, 07 Jun 2024 20:34:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 24.4.4
+
+Fri, 07 Jun 2024 20:34:51 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v29.3.3
 
 ## 24.4.3